+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - `odoo-helper test`: Test only installable addons
 - `odoo-helper addons update-list` command support odoo 7.0
 - `odoo-helper addons test-installed` command support odoo 7.0
+- `odoo-helper fetch`: added experimental support of Mercurial
 
 
 ## Version 0.1.0